Split apart the model code
